capi: Move error functionality into module of same name
Move the error functionality into a separate module to clean up the main
lib.rs file a little bit.
